refactor: extract _merging.py from _sections.py (9 functions, 10 constants)
Move section merging, deduplication, and paragraph-range utilities into
a new _merging.py module. _sections.py drops from 2,538 to 1,955 lines.

Functions moved: _merge_bare_heading_pairs, _merge_adjacent_duplicate_sections,
_strip_printed_toc_page_runs, _drop_empty_interior_title_repeats,
_merge_chapter_subtitle_sections, _merge_chapter_description_paragraphs,
_paragraph_range_between, _has_paragraphs_between,
_paragraphs_between_are_metadata_only.
